Redistribute the new general code to more appropriate locations.
