+++ src/lib/graphics/graphics/Texting.cpp
@@ -88,7 +88,7 @@ void Texting::render() const
 
     // glDrawElementsInstancedBaseInstance not supported on macOS,
     // so workaround using several instance buffers.
-    for (const Pooled<PerTextureVao> & perTexture : mPerTexture)
+    for (const handy::Pooled<PerTextureVao> & perTexture : mPerTexture)
     {
         glBindVertexArray(perTexture->vao);
         bind(*perTexture->texture);
