+++ src/lib/graphics/graphics/TrivialLineStrip.cpp
@@ -48,7 +48,8 @@ TrivialLineStrip::TrivialLineStrip(Size2<int> aRenderResolution) :
         make_VertexSpecification(),
         make_Program()
     },
-    mIbo{ makeLoadedIndexBuffer(gsl::span<Index>{}, BufferHint::StreamDraw) }
+    mIbo{ loadIndexBuffer(mDrawContext.mVertexSpecification.mVertexArray,
+                          gsl::span<Index>{}, BufferHint::StreamDraw) }
 {
     setCameraTransformation(math::AffineMatrix<3, GLfloat>::Identity());
     setProjectionTransformation(math::trans2d::window<GLfloat>(
